+++ packages/holz-logger/vite.config.ts
@@ -1,9 +1,11 @@
 import { readFile } from 'fs/promises';
+import { builtinModules } from 'module';
 import { defineConfig } from 'vite';
 
 export default defineConfig(async () => {
   const pkg = JSON.parse(await readFile('./package.json', 'utf-8'));
-  const externals = Object.keys(pkg.dependencies ?? {});
+  const dependencies = Object.keys(pkg.dependencies ?? {});
+  const builtins = builtinModules.map((name) => `node:${name}`);
 
   return {
     build: {
@@ -14,7 +16,7 @@ export default defineConfig(async () => {
         },
       },
       rollupOptions: {
-        external: externals,
+        external: dependencies.concat(builtins),
         output: {
           exports: 'named',
         },